fix(indexer): SMI-4861 Wave 1 — CI lint fixes + 500-line gate compliance
CI on PR #1089 surfaced two issues on commit 5e9771a8:

1. Prettier format drift in 3 indexer files (lint check failure).
2. Wave 1 changes pushed high-trust-indexer.ts (460→587) and
   skill-processor.ts (498→505) over CLAUDE.md's <500-line gate.
   Pattern previously hit by SMI-4782/4795/4538 retros.

Refactor — extract two cohesive modules out of high-trust-indexer.ts:

  scripts/indexer/tree-hash-cache.ts (55L)
    TreeHashCacheEntry, TreeHashCache, TreeHashCacheCounters,
    TREE_HASH_CACHE_TTL_MS, treeHashCacheKey(), treeHashCacheHit().
    high-trust-indexer.ts re-exports these so existing imports in
    run.ts / phases/high-trust.ts continue to resolve unchanged.

  scripts/indexer/index-skills-from-contents.ts (196L)
    indexSkillsFromContents() + RepoData. The plain-path Contents API
    scan helper. Imported back into high-trust-indexer for the wildcard
    branch's plain-path fallback.

Post-refactor sizes:
  high-trust-indexer.ts: 587 → 345
  skill-processor.ts:    505 → 499 (Wave 1 comment compaction)
  index-skills-from-contents.ts: NEW 196
  tree-hash-cache.ts:    NEW 55

Verbose Wave 1 comments also tightened across high-trust-indexer.ts +
skill-processor.ts to follow CLAUDE.md \"default to no comments\".

Security Audit failure (npm audit high-severity) is the chronic
OpenTelemetry/zod hoist gap — SMI-4642, not blocked on this PR.

Tests: 79/79 indexer pass (incl. 25 Wave 1). Lint+typecheck clean.

[skip-impl-check]
